+++ src/remotes/mod.rs
@@ -1,3 +1,7 @@
+mod macros;
+mod remotecontrol;
+pub use remotecontrol::*;
+
 #[cfg(feature = "nec")]
 pub mod nec;
 #[cfg(feature = "rc5")]
@@ -7,4 +11,5 @@ pub mod rc6;
 #[cfg(feature = "sbp")]
 pub mod sbp;
 
-pub mod logging;
+#[cfg(feature = "std")]
+pub mod std;