+++ CHANGELOG.md
@@ -1,7 +1,12 @@
 # Changelog
+## [Unreleased] - 2019-11-XX
 
-## [Unreleased] - 2019-10-XX
+## [0.5.0] - 2019-11-07
  - Added (optional) embedded-hal support
+ - Added support for NEC commands with 16 bit addresses
+ - Added support for Samsung BluRay Player protocol
+ - Refactored remotes
+ - Cleaned up examples
 
 ## [0.4.1] - 2019-09-29
  - Added support for Rc5 transmit